[#3193] Karl/Phil * Removed ruby instructions as it's not needed anymore * Commented out slack messages * Add jekyll build and serve scripts
This change was done because we had difficulty getting Gemfile +
Gemfile.lock working in the CI environment - due to the need to
download new dependencies at build time (due in part to MITM at the
edge of the VA network which prevented downloading of new gems from
within Jenkins). Thus we decided it was better to avoid that altogether
and just use the exact dependencies as published in the official
Jekyll 4.0 image.